0x14. C - Bit manipulation, 5.101

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned long int aux;
-	unsigned long b;
+	unsigned int b;
 
 	aux = n ^ m;
 	b = 0;
